Resolve CMS quality checks and simplify gallery generation
